+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 
 ### v0.1.0
 - test.static.ts
+- created\_at && updated\_at
+- /test > async\_await 부분 수정